+++ src/core/engine/rendering/lowlevelapi/VertexBuffer.h
@@ -6,17 +6,17 @@
 class VertexBuffer : public RenderApiObject {
 private:
 	static thread_local unsigned int currentlyBoundVBO;
-	int m_size;
+	size_t m_size;
 	
 public:
 	static void bindDefault();
 	static void syncBinding();
 
-	VertexBuffer(const void* data, int size); // size is in bytes
+	VertexBuffer(const void* data, size_t size); // size is in bytes
 	VertexBuffer(VertexBuffer&& other) noexcept;
 	virtual ~VertexBuffer();
 
-	inline int getByteSize() const { return m_size; };
+	inline size_t getByteSize() const { return m_size; };
 
 	void bind() const;
